bug#170: $0 when wallet not connected (#171)
* fixed prices for non metamask user

* fixed prices for non metamask user
