+++ src/store/main/sagas/getBalances.ts
@@ -37,47 +37,6 @@ export function* getBalances(address: string) {
     RICAddress, rexLPETHAddress,
   ];
 
-  const coingeckoIds = new Map<string, string>([
-    [DAIxAddress, 'dai'],
-    [USDCxAddress, 'usd-coin'],
-    [WETHxAddress, 'weth'],
-    [WBTCxAddress, 'wrapped-bitcoin'],
-    [MATICxAddress, 'matic-network'],
-    [MKRxAddress, 'maker'],
-  ]);
-
-  async function getCoingeckoRates() {
-    const ids = [...coingeckoIds.values()];
-    const coingeckoRequestUrl =
-          `https://api.coingecko.com/api/v3/coins/markets?vs_currency=USD&ids=${ids.join(',')}`;
-    return fetch(coingeckoRequestUrl, {
-      method: 'GET',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-    })
-      .then((response) => response.json())
-      .then((responseData) => responseData)
-      .catch((error) => console.error(error));
-  }
-
-  const tokenAddresses = [...coingeckoIds.keys()];
-  const coingeckoPrices: { [key:string]: number } = {};
-
-  getCoingeckoRates().then((response:any) => {
-    tokenAddresses.forEach((tokenAddress) => {
-      const id = coingeckoIds?.get(tokenAddress);
-      const tokenData = response.filter((res:any) => res.id === id!);
-      if (tokenData === undefined) {
-        coingeckoPrices[tokenAddress] = 0;
-        console.warn('Could not fetch price for token ', tokenAddress);
-      } else {
-        coingeckoPrices[tokenAddress] = tokenData[0].current_price;
-      }
-    });
-  });
-
   const main: ReturnType<typeof selectMain> = yield select(selectMain);
   const { web3 } = main;
   const requests = contractsAddress.map((el) => ({
@@ -103,5 +62,5 @@ export function* getBalances(address: string) {
   // Edit matic balance
   balances[WMATICAddress] = fromWei(yield web3.eth.getBalance(address), 18);
 
-  yield put(mainSetState({ balances, coingeckoPrices }));
+  yield put(mainSetState({ balances }));
 }
