+++ src/store/main/sagas/loadData.ts
@@ -17,6 +17,7 @@ import {
 import { getBalances } from './getBalances';
 import { sweepQueryFlow } from './sweepQueryFlow';
 import { selectMain } from '../selectors';
+import { getCoingeckoPrices } from '../../../utils/getCoingeckoPrices';
 
 export function* loadData() {
   try {
@@ -24,6 +25,7 @@ export function* loadData() {
     const main: ReturnType<typeof selectMain> = yield select(selectMain);
     const { web3 } = main;
     const address: Unwrap<typeof getAddress> = yield call(getAddress, web3);
+    const coingeckoPrices: Unwrap<typeof getCoingeckoPrices> = yield call(getCoingeckoPrices);
     yield call(getBalances, address);
     yield all([
       call(checkIfApproveUsdc),
@@ -38,6 +40,7 @@ export function* loadData() {
     ]);
     yield put(mainSetState({
       address,
+      coingeckoPrices,
       isLoading: false,
     }));
   } catch (e) {
